Projet/requirements.txt
pandas
requests
scikit-learn
paho-mqtt
pytest
